Feat: Updated DB Documentation

+++ src/main/resources/db/migration/V2__Create_products_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS  products (
                           id SERIAL PRIMARY KEY,
-                          farmer_id BIGINT NOT NULL,
+                          farmer_id INT NOT NULL,
                           name VARCHAR(255) NOT NULL,
                           category ENUM('vegetable', 'fruit', 'grain', 'dairy', 'meat') NOT NULL,
                           price DECIMAL(10,2) NOT NULL,
